pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fpt-mcp"
version = "0.1.0"
description = "MCP server for Autodesk Flow Production Tracking (ShotGrid)"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "mcp[cli]>=1.0.0",
    "shotgun-api3>=3.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
fpt-mcp = "fpt_mcp.server:main"
